requirements.txt
Babel==2.3.4
Django==1.10.1
django-allauth==0.27.0
django-phonenumber-field==1.1.0
gunicorn==19.6.0
oauthlib==2.0.0
phonenumberslite==7.5.2
Pillow==3.3.1
psycopg2==2.6.2
python3-openid==3.0.10
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
whitenoise==3.2.1
